Implemented new validations in the RecipeRepositoryImpl class

+++ src/main/java/com/zord/recipe/api/repositories/RecipeRepositoryImpl.java
@@ -3,6 +3,7 @@ package com.zord.recipe.api.repositories;
 import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoDatabase;
 import com.mongodb.client.model.Filters;
+import com.zord.recipe.api.exceptions.ExistsUserIdException;
 import com.zord.recipe.api.exceptions.ObjectNotFoundException;
 import com.zord.recipe.api.model.Comment;
 import com.zord.recipe.api.model.Recipe;
@@ -12,6 +13,7 @@ import org.bson.types.ObjectId;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
 
 import static com.mongodb.client.model.Filters.eq;
 import static com.mongodb.client.model.Filters.regex;
@@ -29,6 +31,9 @@ public class RecipeRepositoryImpl implements RecipeRepository {
         List<Recipe> result = new ArrayList<>();
         for (Recipe recipe : coll.find()) {
             result.add(recipe);
+        }
+        if (result.isEmpty()) {
+            throw new ObjectNotFoundException("Nenhuma receita encontada!");
         }
         return result;
     }
@@ -49,6 +54,9 @@ public class RecipeRepositoryImpl implements RecipeRepository {
         for (Recipe recipe : coll.find().filter(regex("ingredients", ingredient))) {
             result.add(recipe);
         }
+        if (result.isEmpty()) {
+            throw new ObjectNotFoundException("Nenhuma receita encontada!");
+        }
         Collections.sort(result);
         return result;
     }
@@ -63,6 +71,9 @@ public class RecipeRepositoryImpl implements RecipeRepository {
             if (!result.contains(recipe)) {
                 result.add(recipe);
             }
+        }
+        if (result.isEmpty()) {
+            throw new ObjectNotFoundException("Nenhuma receita encontada!");
         }
         Collections.sort(result);
         return result;
@@ -79,18 +90,28 @@ public class RecipeRepositoryImpl implements RecipeRepository {
     @Override
     public Recipe update(String id, Recipe recipe) {
         Bson filter = eq("_id", id);
-        return coll.findOneAndReplace(filter, recipe);
+        Recipe updatedRecipe = coll.findOneAndReplace(filter, recipe);
+        if (updatedRecipe == null) {
+            throw new ObjectNotFoundException("Não foi possível atualizar a receita. Id: " + id + " inexistente!");
+        }
+        return updatedRecipe;
     }
 
     @Override
     public void delete(String id) {
         Bson filter = Filters.eq("_id", id);
+        if (coll.find(filter).first() == null) {
+            throw new ObjectNotFoundException("Não foi possível deletar a receita. Id: " + id + " inexistente!");
+        }
         coll.deleteOne(filter);
     }
 
     @Override
     public Recipe addLike(Integer userId, String recipeId) {
         Recipe recipe = findById(recipeId);
+        if (recipe.getLikes().contains(userId)) {
+            throw new ExistsUserIdException("Usuário id: " + userId + " já curtiu a receita id: " + recipeId + "!");
+        }
         recipe.getLikes().add(userId);
         update(recipeId, recipe);
         return findById(recipeId);
@@ -99,6 +120,9 @@ public class RecipeRepositoryImpl implements RecipeRepository {
     @Override
     public void removeLike(Integer userId, String recipeId) {
         Recipe recipe = findById(recipeId);
+        if (!recipe.getLikes().contains(userId)) {
+            throw new ObjectNotFoundException("O usuário id: " + userId + " não curtiu a receita id: " + recipeId + "!");
+        }
         recipe.getLikes().remove(userId);
         update(recipeId, recipe);
     }
@@ -120,7 +144,20 @@ public class RecipeRepositoryImpl implements RecipeRepository {
     @Override
     public void removeComment(String recipeId, String commentId) {
         Recipe recipe = findById(recipeId);
+        if (!commentExists(recipe, commentId)) {
+            throw new ObjectNotFoundException("Comentário id: " + commentId + " não encontrado!");
+        }
         recipe.getComments().removeIf(c -> c.getId().equals(commentId));
         update(recipeId, recipe);
+    }
+
+    private boolean commentExists(Recipe recipe, String commentId) {
+       List<Comment> comments = recipe.getComments();
+       for (Comment c : comments) {
+           if (c.getId().equals(commentId)) {
+               return true;
+           }
+       }
+       return false;
     }
 }
